Implmented the safer format function.
This implmentation isn't particularly fast, but it should not be too slow either
because the logic is quite straightforward. The only issue is that characters
are printed one by one when in fact it would be safe to assume that there are
only few occurence of the format character in the provided template.